requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pickle-mixin
transformers
torch
